Fix browser runtime disconnect cleanup
